requirements.txt
streamlit
boto3
langchain
langchain-community
faiss-cpu
numpy
pypdf